+++ src/test/java/dk/kb/present/config/ServiceConfigTest.java
@@ -5,9 +5,7 @@ import dk.kb.util.yaml.YAML;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;