+++ src/test/java/dk/kb/present/transform/XSLTCumulusToSchemaDotOrgTransformerTest.java
@@ -7,7 +7,6 @@ import com.google.gson.JsonParser;
 import dk.kb.present.TestFiles;
 import dk.kb.present.TestUtil;
 import dk.kb.present.config.ServiceConfig;
-import dk.kb.util.Resolver;
 import dk.kb.util.yaml.YAML;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;