+++ src/test/java/dk/kb/present/transform/XSLTCumulusToSolrTransformerTest.java
@@ -18,7 +18,6 @@ import dk.kb.present.TestFiles;
 import dk.kb.present.TestUtil;
 
 import dk.kb.present.config.ServiceConfig;
-import dk.kb.util.Resolver;
 import dk.kb.util.yaml.YAML;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;