+++ src/test/java/dk/kb/present/transform/XSLTTransformerTestBase.java
@@ -17,7 +17,6 @@ package dk.kb.present.transform;
 import dk.kb.present.TestUtil;
 import dk.kb.present.config.ServiceConfig;
 import dk.kb.present.util.TestFileProvider;
-import dk.kb.util.Resolver;
 import dk.kb.util.yaml.YAML;
 import org.junit.jupiter.api.BeforeAll;
 import org.slf4j.Logger;